EDM-119: Allow decimals for  hours_worked_per_week (#2405)
* Allow partial hours worked per week

* Update openapi.yaml

* Fix lint, build

* Revert make format sort changes

* Update tests

* Update tests

* Fix lint

+++ api/massgov/pfml/api/models/claims/responses.py
@@ -1,4 +1,5 @@
 from datetime import date
+from decimal import Decimal
 from typing import List, Optional
 
 from massgov.pfml.api.models.claims.common import (
@@ -22,7 +23,7 @@ class ClaimReviewResponse(PydanticBaseModel):
     employer_fein: Optional[FEINFormattedStr]
     fineos_absence_id: Optional[str]
     first_name: Optional[str]
-    hours_worked_per_week: Optional[int]
+    hours_worked_per_week: Optional[Decimal]
     last_name: Optional[str]
     leave_details: Optional[LeaveDetails]
     middle_name: Optional[str]
